requirements.txt
numpy
streamlit==1.46.1
tensorflow-cpu==2.20.0
Pillow<12
